+++ __tests__/test.test.js
@@ -1,239 +1,7 @@
-import { ERROR_MESSAGE, RANK } from "../src/constants";
-import InputView from "../src/View/InputView";
+import { RANK } from "../src/constants";
 import Lotto from "../src/Model/Lotto";
-import LottoGenerator from "../src/LottoGenerator";
 import ScoreBoard from "../src/ScoreBoard";
-import Utils from "../src/Utils";
-import Validator from "../src/Validator";
 import WinningLotto from "../src/Model/WinningLotto";
-
-const mockQuestions = (inputs) => {
-  Utils.readLineAsync = jest.fn();
-
-  Utils.readLineAsync.mockImplementation(() => {
-    const input = inputs.shift();
-
-    return Promise.resolve(input);
-  });
-};
-
-describe("로또 구입 금액 입력 테스트", () => {
-  test("빈 값이 입력되는 경우 에러를 발생 시켜야 한다 ", async () => {
-    // given
-    const input = [""];
-    mockQuestions(input);
-
-    // when & then
-    await expect(InputView.readMoney("")).rejects.toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("문자가 포함되는 경우 에러를 발생 시켜야 한다 ", async () => {
-    // given
-    const input = ["1000j"];
-    mockQuestions(input);
-
-    // when & then
-    await expect(InputView.readMoney()).rejects.toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("1000원 단위 입력이 아닌 경우 에러를 발생 시켜야 한다 ", async () => {
-    // given
-    const input = ["1500"];
-    mockQuestions(input);
-
-    // when & then
-    await expect(InputView.readMoney()).rejects.toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("입력이 0인 경우 에러를 발생 시켜야 한다 ", async () => {
-    // given
-    const input = ["0"];
-    mockQuestions(input);
-
-    // when & then
-    await expect(InputView.readMoney()).rejects.toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("입력이 음수인 경우 에러를 발생 시켜야 한다 ", async () => {
-    // given
-    const input = ["-1"];
-    mockQuestions(input);
-
-    // when & then
-    await expect(InputView.readMoney()).rejects.toThrow(ERROR_MESSAGE.PREFIX);
-  });
-});
-
-describe("로또 발행 테스트", () => {
-  test("구매한 로또의 개수를 올바르게 계산해야 한다.", () => {
-    // given
-    const money = 10_000;
-
-    // when & then
-    expect(LottoGenerator.calculateBuyLottoCount(money)).toEqual(10);
-  });
-
-  test("로또를 생성한다.", () => {
-    // given
-    const lotto = new Lotto([6, 5, 4, 3, 2, 1]);
-
-    // when & then
-    expect(lotto).toBeInstanceOf(Lotto);
-  });
-
-  test("0이 입력된 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongNumbers = [0, 1, 2, 3, 4, 5];
-    // when & then
-    expect(() => new Lotto(wrongNumbers)).toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("음의 정수가 입력된 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongNumbers = [-1, 1, 2, 3, 4, 5];
-    // when & then
-    expect(() => new Lotto(wrongNumbers)).toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("1~45 사이가 아닌 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongNumbers = [1, 2, 3, 4, 5, 100];
-    // when & then
-    expect(() => new Lotto(wrongNumbers)).toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("번호가 중복되는 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongNumbers = [1, 1, 2, 3, 4, 5];
-    // when & then
-    expect(() => new Lotto(wrongNumbers)).toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("로또 번호가 6개가 아닌 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongNumbers = [1, 2, 3, 4, 5];
-    // when & then
-    expect(() => new Lotto(wrongNumbers)).toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("로또 생성 시 번호를 오름차순으로 올바르게 정렬해야 한다", () => {
-    // given
-    const unSortedNunmbers = [6, 5, 4, 3, 2, 1];
-
-    // when
-    const lotto = new Lotto(unSortedNunmbers);
-
-    // then
-    expect(lotto.getNumbers()).toEqual([1, 2, 3, 4, 5, 6]);
-  });
-
-  test("구입 수량만큼 로또를 발행해야 한다", () => {
-    // given
-    const buyLottoCount = 10;
-
-    // when
-    const lottos = LottoGenerator.makeLottos(buyLottoCount);
-
-    // then
-    expect(lottos.length).toEqual(buyLottoCount);
-  });
-});
-
-describe("유틸 함수 테스트", () => {
-  test("1~45 사이의 무작위 숫자 하나를 반환해야 한다", () => {
-    // given & when
-    const randomNumber = Utils.getRandomNumber(1, 45);
-
-    // then 나중에 모킹 해놓기
-    expect(randomNumber >= 1 && randomNumber <= 45).toEqual(true);
-  });
-
-  test("1~45 사이의 중복되지 않는 무작위 숫자 6개를 반환해야 한다", () => {
-    // given
-    const count = 6;
-
-    // when
-    const randomNumbers = Utils.getRandomNumbers(1, 45, count);
-
-    // then 나중에 모킹 해놓기
-    expect(randomNumbers.length).toEqual(count);
-    expect(randomNumbers.length === new Set(randomNumbers).size).toEqual(true);
-  });
-});
-
-describe("당첨 로또 테스트", () => {
-  test("당첨 로또를 생성한다.", () => {
-    // given
-    const winningNumbers = [1, 2, 3, 4, 5, 6];
-    const bonusNumber = 7;
-
-    // when
-    const winningLotto = new WinningLotto(winningNumbers, bonusNumber);
-
-    // then
-    expect(winningLotto).toBeInstanceOf(WinningLotto);
-  });
-
-  test("당첨 번호에 0이 포함된 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongWinningNumbers = [0, 1, 2, 3, 4, 5];
-    const bonusNumber = 7;
-    // when & then
-    expect(() => new WinningLotto(wrongWinningNumbers, bonusNumber)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("당첨 번호에 음의 정수가 포함된 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongWinningNumbers = [-1, 1, 2, 3, 4, 5];
-    const bonusNumber = 7;
-    // when & then
-    expect(() => new WinningLotto(wrongWinningNumbers, bonusNumber)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("당첨 번호에 1~45 사이가 아닌 값이 포함된 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongWinningNumbers = [1, 2, 3, 4, 5, 100];
-    const bonusNumber = 7;
-    // when & then
-    expect(() => new WinningLotto(wrongWinningNumbers, bonusNumber)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("당첨 번호가 중복되는 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongWinningNumbers = [1, 1, 2, 3, 4, 5];
-    const bonusNumber = 7;
-    // when & then
-    expect(() => new WinningLotto(wrongWinningNumbers, bonusNumber)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("당첨 번호가 6개가 아닌 경우 에러를 발생시킨다", () => {
-    // given
-    const wrongWinningNumbers = [1, 2, 3, 4, 5];
-    const bonusNumber = 7;
-    // when & then
-    expect(() => new WinningLotto(wrongWinningNumbers, bonusNumber)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("보너스 번호가 당첨 번호와 중복되는 경우 에러를 발생시킨다", () => {
-    // given
-    const winningNUmbers = [1, 2, 3, 4, 5, 6];
-    const wrongBonusNumber = 1;
-    // when & then
-    expect(() => new WinningLotto(winningNUmbers, wrongBonusNumber)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-});
 
 describe("당첨 여부 테스트", () => {
   test("구매한 로또 하나와 당첨로또를 비교해서 일치하는 번호의 개수를 반환해야한다", () => {
@@ -267,7 +35,7 @@ describe("당첨 여부 테스트", () => {
     [RANK.FOURTH.DISPLAY, RANK.FOURTH.MATCH_COUNT, true],
     [RANK.FIFTH.DISPLAY, RANK.FIFTH.MATCH_COUNT, true],
   ])(
-    "일치하는 번호 개수와 보너스 번호의 개수를 확인해 올바르게 등수를 계산해야 한다.(%s등)",
+    "일치하는 번호 개수와 보너스 번호의 개수를 확인해 올바르게 등수를 계산해야 한다.(%s)",
     (expectedRank, matchCount, hasBonus) => {
       // when
       const rank = ScoreBoard.getRank(matchCount, hasBonus);
@@ -316,57 +84,5 @@ describe("당첨 여부 테스트", () => {
 
     // then
     expect(profitRate).toEqual(result.toFixed(1));
-  });
-});
-
-describe("유효성 검증 테스트", () => {
-  test("입력된 문자열이 빈 값이면 에러를 발생시켜야 한다", () => {
-    // given
-    const wrongInput = "";
-
-    // when & then
-    expect(() => Validator.validateNotEmptyString(wrongInput)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-  test("입력된 문자열이 숫자가 아니면 에러를 발생시켜야 한다", () => {
-    // given
-    const wrongInput = "NoNumber";
-
-    // when & then
-    expect(() => Validator.validateStringIsNumber(wrongInput)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("숫자가 특정 값으로 나누어 떨어지지 않으면 에러를 발생시켜야 한다", () => {
-    // given
-    const divideNumber = 1000;
-    const wrongInput = 1500;
-
-    // when & then
-    expect(() =>
-      Validator.validateNumberDivided(wrongInput, divideNumber),
-    ).toThrow(ERROR_MESSAGE.PREFIX);
-  });
-
-  test("배열에 중복된 원소가 존재하면 에러를 발생시켜야 한다", () => {
-    // given
-    const wrongArray = [1, 1, 2, 3, 4, 5];
-
-    // when & then
-    expect(() => Validator.validateNotDuplicated(wrongArray)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
-  });
-
-  test("배열의 길이가 올바르지 않으면 에러를 발생시켜야 한다", () => {
-    // given
-    const wrongArray = [1, 2, 3, 4, 5];
-
-    // when & then
-    expect(() => Validator.validateArrayLength(wrongArray, 6)).toThrow(
-      ERROR_MESSAGE.PREFIX,
-    );
   });
 });
